sysPrxForUser module split
