+++ rpcs3/Emu/Cell/RawSPUThread.h
@@ -1,11 +1,12 @@
 #pragma once
+
 #include "SPUThread.h"
 
 enum : u32
 {
-	RAW_SPU_OFFSET = 0x00100000,
-	RAW_SPU_BASE_ADDR = 0xE0000000,
-	RAW_SPU_LS_OFFSET = 0x00000000,
+	RAW_SPU_OFFSET      = 0x00100000,
+	RAW_SPU_BASE_ADDR   = 0xE0000000,
+	RAW_SPU_LS_OFFSET   = 0x00000000,
 	RAW_SPU_PROB_OFFSET = 0x00040000,
 };
