+++ rpcs3/Emu/SysCalls/Modules/sys_mmapper_.cpp
@@ -1,0 +1,19 @@
+#include "stdafx.h"
+#include "Emu/Memory/Memory.h"
+#include "Emu/System.h"
+#include "Emu/SysCalls/Modules.h"
+
+#include "Emu/SysCalls/lv2/sys_mmapper.h"
+#include "sysPrxForUser.h"
+
+extern Module sysPrxForUser;
+
+void sysPrxForUser_sys_mmapper_init()
+{
+	// TODO: split syscalls and liblv2 functions
+	REG_FUNC(sysPrxForUser, sys_mmapper_allocate_memory);
+	REG_FUNC(sysPrxForUser, sys_mmapper_allocate_memory_from_container);
+	REG_FUNC(sysPrxForUser, sys_mmapper_map_memory);
+	REG_FUNC(sysPrxForUser, sys_mmapper_unmap_memory);
+	REG_FUNC(sysPrxForUser, sys_mmapper_free_memory);
+}
