+++ rpcs3/Emu/SysCalls/lv2/sys_process.cpp
@@ -55,156 +55,6 @@ s32 sys_process_exit(s32 status)
 	return CELL_OK;
 }
 
-void sys_game_process_exitspawn(vm::cptr<char> path, u32 argv_addr, u32 envp_addr, u32 data_addr, u32 data_size, u32 prio, u64 flags)
-{
-	std::string _path = path.get_ptr();
-	const std::string& from = "//";
-	const std::string& to = "/";
-
-	size_t start_pos = 0;
-	while ((start_pos = _path.find(from, start_pos)) != std::string::npos) {
-		_path.replace(start_pos, from.length(), to);
-		start_pos += to.length();
-	}
-
-	sys_process.Todo("sys_game_process_exitspawn()");
-	sys_process.Warning("path: %s", _path.c_str());
-	sys_process.Warning("argv: 0x%x", argv_addr);
-	sys_process.Warning("envp: 0x%x", envp_addr);
-	sys_process.Warning("data: 0x%x", data_addr);
-	sys_process.Warning("data_size: 0x%x", data_size);
-	sys_process.Warning("prio: %d", prio);
-	sys_process.Warning("flags: %d", flags);
-
-	std::vector<std::string> argv;
-	std::vector<std::string> env;
-
-	if (argv_addr)
-	{
-		auto argvp = vm::cpptr<char>::make(argv_addr);
-		while (argvp && *argvp)
-		{
-			argv.push_back(argvp[0].get_ptr());
-			argvp++;
-		}
-
-		for (auto &arg : argv) {
-			sys_process.Log("argument: %s", arg.c_str());
-		}
-	}
-
-	if (envp_addr)
-	{
-		auto envp = vm::cpptr<char>::make(envp_addr);
-		while (envp && *envp)
-		{
-			env.push_back(envp[0].get_ptr());
-			envp++;
-		}
-
-		for (auto &en : env) {
-			sys_process.Log("env_argument: %s", en.c_str());
-		}
-	}
-
-	//TODO: execute the file in <path> with the args in argv
-	//and the environment parameters in envp and copy the data
-	//from data_addr into the adress space of the new process
-	//then kill the current process
-
-	Emu.Pause();
-	sys_process.Success("Process finished");
-
-	CallAfter([]()
-	{
-		Emu.Stop();
-	});
-
-	std::string real_path;
-
-	Emu.GetVFS().GetDevice(_path.c_str(), real_path);
-
-	Emu.BootGame(real_path, true);
-
-	return;
-}
-
-void sys_game_process_exitspawn2(vm::cptr<char> path, u32 argv_addr, u32 envp_addr, u32 data_addr, u32 data_size, u32 prio, u64 flags)
-{
-	std::string _path = path.get_ptr();
-	const std::string& from = "//";
-	const std::string& to = "/";
-
-	size_t start_pos = 0;
-	while ((start_pos = _path.find(from, start_pos)) != std::string::npos) {
-		_path.replace(start_pos, from.length(), to);
-		start_pos += to.length();
-	}
-
-	sys_process.Warning("sys_game_process_exitspawn2()");
-	sys_process.Warning("path: %s", _path.c_str());
-	sys_process.Warning("argv: 0x%x", argv_addr);
-	sys_process.Warning("envp: 0x%x", envp_addr);
-	sys_process.Warning("data: 0x%x", data_addr);
-	sys_process.Warning("data_size: 0x%x", data_size);
-	sys_process.Warning("prio: %d", prio);
-	sys_process.Warning("flags: %d", flags);
-
-	std::vector<std::string> argv;
-	std::vector<std::string> env;
-
-	if (argv_addr)
-	{
-		auto argvp = vm::cpptr<char>::make(argv_addr);
-		while (argvp && *argvp)
-		{
-			argv.push_back(argvp[0].get_ptr());
-			argvp++;
-		}
-
-		for (auto &arg : argv)
-		{
-			sys_process.Log("argument: %s", arg.c_str());
-		}
-	}
-
-	if (envp_addr)
-	{
-		auto envp = vm::cpptr<char>::make(envp_addr);
-		while (envp && *envp)
-		{
-			env.push_back(envp[0].get_ptr());
-			envp++;
-		}
-
-		for (auto &en : env)
-		{
-			sys_process.Log("env_argument: %s", en.c_str());
-		}
-	}
-
-	//TODO: execute the file in <path> with the args in argv
-	//and the environment parameters in envp and copy the data
-	//from data_addr into the adress space of the new process
-	//then kill the current process
-
-	Emu.Pause();
-	sys_process.Success("Process finished");
-
-	CallAfter([]()
-	{
-		Emu.Stop();
-	});
-	
-	std::string real_path;
-
-	Emu.GetVFS().GetDevice(_path.c_str(), real_path);
-
-	Emu.BootGame(real_path, true);
-
-	return;
-}
-
 s32 sys_process_get_number_of_object(u32 object, vm::ptr<u32> nump)
 {
 	sys_process.Error("sys_process_get_number_of_object(object=0x%x, nump=*0x%x)", object, nump);
