+++ rpcs3/emucore.vcxproj
@@ -297,11 +297,21 @@
     <ClCompile Include="Emu\SysCalls\Modules\sceNpUtil.cpp" />
     <ClCompile Include="Emu\SysCalls\Modules\cellVideoUpload.cpp" />
     <ClCompile Include="Emu\SysCalls\Modules\sysPrxForUser.cpp" />
+    <ClCompile Include="Emu\SysCalls\Modules\sys_game.cpp" />
+    <ClCompile Include="Emu\SysCalls\Modules\sys_heap.cpp" />
     <ClCompile Include="Emu\SysCalls\Modules\sys_lv2dbg.cpp" />
     <ClCompile Include="Emu\SysCalls\Modules\cellHttp.cpp" />
     <ClCompile Include="Emu\SysCalls\Modules\sys_io.cpp" />
     <ClCompile Include="Emu\SysCalls\Modules\sys_libc.cpp" />
+    <ClCompile Include="Emu\SysCalls\Modules\sys_lwcond_.cpp" />
+    <ClCompile Include="Emu\SysCalls\Modules\sys_lwmutex_.cpp" />
+    <ClCompile Include="Emu\SysCalls\Modules\sys_mempool.cpp" />
+    <ClCompile Include="Emu\SysCalls\Modules\sys_mmapper_.cpp" />
     <ClCompile Include="Emu\SysCalls\Modules\sys_net.cpp" />
+    <ClCompile Include="Emu\SysCalls\Modules\sys_ppu_thread_.cpp" />
+    <ClCompile Include="Emu\SysCalls\Modules\sys_prx_.cpp" />
+    <ClCompile Include="Emu\SysCalls\Modules\sys_spinlock.cpp" />
+    <ClCompile Include="Emu\SysCalls\Modules\sys_spu_.cpp" />
     <ClCompile Include="Emu\SysCalls\SysCalls.cpp" />
     <ClCompile Include="Emu\System.cpp" />
     <ClCompile Include="Ini.cpp" />
